Move file to utils folder
