+++ frontend/apps/core-app/src/utils/__test__/buildDefaultRecord.test.ts
