+++ frontend/apps/core-app/src/utils/buildDefaultRecord.ts
